% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{add_in_taxa}
\alias{add_in_taxa}
\title{Adds in taxa if silva database
Returns MetaScope Table with silva taxa in separate columns}
\usage{
add_in_taxa(combined_pre, caching, path_to_write)
}
\arguments{
\item{combined_pre}{MetaScope ID file with silva taxa}

\item{caching}{Boolean for if all_silva_headers.rds is already downloaded}

\item{path_to_write}{Path to save all_silva_headers.rds}
}
\description{
Adds in taxa if silva database
Returns MetaScope Table with silva taxa in separate columns
}
\keyword{internal}
