% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{align_details}
\alias{align_details}
\title{A universal parameter settings object for Rsubread alignment}
\format{
An object of class \code{list} of length 7.
}
\usage{
align_details
}
\description{
This object is a named vector of multiple options that can be chosen for
functions that involve alignment with Rsubread, namely `align_target()`
and `filter_host()`. These functions take an object for the parameter
`settings`, which are provided by `align_details` by default, or may
be given by a user-created object containing the same information.
}
\details{
The default options included in `align_details` are `type = "dna"`,
`nthreads = "8`, `maxMismatches = 5`, `nsubreads = 10`, `phredOffset = 33`,
`unique = FALSE`, and `nBestLocations = 16`.
Full descriptions of these parameters can be read by
acessing `?Rsubread::align`.
}
\examples{
data("align_details")

}
\keyword{datasets}
