% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_target.R
\name{align_target_bowtie}
\alias{align_target_bowtie}
\title{Align microbiome reads to set of indexed Bowtie2 libraries}
\usage{
align_target_bowtie(
  read1,
  read2 = NULL,
  index_dir,
  index_basename,
  indexFormat = "s",
  sam_dir,
  sam_basename = basename(read1),
  threads = 4,
  overwrite = FALSE
)
}
\arguments{
\item{read1}{Location of the .fastq file to align.}

\item{read2}{Optional. Location of the .fastq file to align.}

\item{index_dir}{Directory location that contains the bowtie2 indexes.}

\item{index_basename}{Basename of the bowtie2 indexes.}

\item{sam_dir}{Directory location where the output file should be created.}

\item{sam_basename}{Basename of the output file.}

\item{threads}{The number of threads that should be utilized}

\item{overwrite}{Whether existing files should be overwritten}
}
\description{
Align microbiome reads to set of indexed Bowtie2 libraries
}
