% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{blastn_results}
\alias{blastn_results}
\title{rBlast_results}
\usage{
blastn_results(
  results_table,
  bam_file,
  num_results = 10,
  num_reads_per_result = 100,
  hit_list = 10,
  num_threads = 1,
  db_path,
  out_path,
  sample_name = NULL,
  quiet = quiet,
  accessions_path,
  fasta_dir = NULL,
  NCBI_key = NULL
)
}
\arguments{
\item{results_table}{A data.frame of the Metascope results}

\item{bam_file}{A sorted bam file and index file, loaded with
Rsamtools::bamFile}

\item{num_results}{A number indicating number of Metascope results to blast}

\item{num_reads_per_result}{A number indicating number of reads to blast per
result}

\item{hit_list}{A number of how many blast results to fetch for each read}

\item{num_threads}{Number of threads if multithreading}

\item{db_path}{Blast database path}

\item{out_path}{Output directory to save csv files, including base name of
files}

\item{sample_name}{Character string, sample name for output files.}

\item{quiet}{Logical, whether to print out more informative messages. Default
is FALSE.}

\item{accessions_path}{Directory where accession files for blast are stored.}

\item{fasta_dir}{inc.}

\item{NCBI_key}{(character) NCBI Entrez API key. optional. See
taxize::use_entrez(). Due to the high number of requests made to NCBI, this
function will be less prone to errors if you obtain an NCBI key.}
}
\value{
Creates and exports num_results number of csv files with blast
  results from local blast
}
\description{
rBlast_results
}
