% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt2_16S_params}
\alias{bt2_16S_params}
\title{A universal parameter object for Bowtie 2 16S alignment}
\format{
list
}
\usage{
data(bt2_16S_params)
}
\description{
This character string provides several Bowtie 2 options to provide an optimized
alignment specifically optimized for 16S amplicon sequencing data. This object
can be used with functions that use the Bowtie 2 aligner through
the \code{Rbowtie2} package, namely \code{align_target_bowtie()} and
\code{filter_host_bowtie}. These settings can be substituted for default
settings by passing \code{} to the \code{bowtie2_options} argument.
}
\details{
The default parameters listed in this object are
"-local -R 2 -N 0 -L 25 -i S,1,0.75 -k 10 --score-min L,100,1.28".

Further delineation of Bowtie 2 parameters is provided in the
\href{https://bowtie-bio.sourceforge.net/bowtie2/manual.shtml}{Bowtie 2 manual}.

These parameters were chosen from the article
\href{https://doi.org/10.1101/2022.07.27.501757}{Metagenomic profiling pipelines improve taxonomic classification for 16S amplicon sequencing data},
Faits et al. [Preprint].
}
\examples{
data("bt2_16S_params")

}
\keyword{datasets}
