% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt2_loose_params}
\alias{bt2_loose_params}
\title{A universal parameter object for Bowtie 2 loose alignment}
\format{
list
}
\usage{
data(bt2_loose_params)
}
\description{
This character string provides several Bowtie 2 options to provide a loose
alignment useful for metagenomes. This object can be used with functions that
use the Bowtie 2 aligner through the \code{Rbowtie2} package, namely
\code{align_target_bowtie()} and \code{filter_host_bowtie}. These settings
can be substituted for default settings by passing \code{} to the
\code{bowtie2_options} argument.
}
\details{
The default parameters listed in this object are "--local -k 100 -D 20 -R 3
-L 3 -N 1 -p 8 --gbar 1 --mp 3".

Further delineation of Bowtie 2 parameters is provided in the
\href{https://bowtie-bio.sourceforge.net/bowtie2/manual.shtml}{Bowtie 2
manual}.

These parameters were chosen from the article
\href{https://doi.org/10.1101/2022.07.27.501757}{bowtie2: Relaxed Parameters
for Generous Alignments to Metagenomes}.
}
\examples{
data("bt2_loose_params")

}
\keyword{datasets}
