% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{check_blastn_exists}
\alias{check_blastn_exists}
\title{Check if blastn exists on the system}
\usage{
check_blastn_exists()
}
\value{
Returns TRUE if blastn exists on the system, else FALSE.
}
\description{
This is an internal function that is not meant to be used
  outside of the package. It checks whether blastn exists on the system.
}
\details{
Checks if blastn is installed
}
