% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_samtools_exists.R
\name{check_samtools_exists}
\alias{check_samtools_exists}
\title{Check if samtools exists on the system}
\usage{
check_samtools_exists()
}
\value{
Returns TRUE if samtools exists on the system, else FALSE.
}
\description{
This is an internal function that is not meant to be used
  outside of the package. It checks whether samtools exists on the system.
}
