% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_animalcules_silva.R
\name{convert_animalcules_silva}
\alias{convert_animalcules_silva}
\title{Create a multi-assay experiment from MetaScope output for usage with
animalcules with the SILVA database}
\usage{
convert_animalcules_silva(
  meta_counts,
  annot_path,
  which_annot_col,
  end_string = ".metascope_id.csv",
  qiime_biom_out = FALSE,
  path_to_write = "."
)
}
\arguments{
\item{meta_counts}{A vector of filepaths to the counts ID CSVs output by
\code{metascope_id()} created with the SILVA database.}

\item{annot_path}{The filepath to the CSV annotation file for the samples.}

\item{which_annot_col}{The name of the column of the annotation file
containing the sample IDs. These should be the same as the
\code{meta_counts} root filenames.}

\item{end_string}{The end string used at the end of the metascope_id files.
Default is ".metascope_id.csv".}

\item{qiime_biom_out}{Would you also like a qiime-compatible biom file
output? If yes, two files will be saved: one is a biom file of the counts
table, and the other is a specifically formatted mapping file of metadata
information. Default is \code{FALSE}.}

\item{path_to_write}{If \code{qiime_biom_out = TRUE}, where should output QIIME
files be written? Should be a character string of the folder path. Default is
'.', i.e. the current working directory.}
}
\value{
Returns a MultiAssay Experiment file of combined sample counts data
  and/or biom file and mapping file for analysis with QIIME. The MultiAssay
  Experiment will have a counts assay ("MGX").
}
\description{
Upon completion of the MetaScope pipeline, users can analyze and visualize
abundances in their samples using the animalcules package. This function
allows interoperability of \code{metascope_id} output with both animalcules
and QIIME. After running this function, the user should save the returned MAE
to an RDS file using a function like \code{saveRDS} to upload the output into
the \code{animalcules} package. NOTE: This function is for outputs that were
generated with the SILVA database.
}
