% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_accessions.R
\name{download_accessions}
\alias{download_accessions}
\title{Download indexes required for MetaScope ID and MetaBlast}
\usage{
download_accessions(
  ind_dir,
  NCBI_accessions_database = TRUE,
  NCBI_accessions_name = "accessionTaxa.sql",
  silva_taxonomy_database = TRUE,
  silva_taxonomy_name = "all_silva_headers.rds",
  blast_16S_database = TRUE,
  blast_16S_name = "16S_ribosomal_RNA"
)
}
\arguments{
\item{ind_dir}{Character path to directory where indices should be saved}

\item{NCBI_accessions_database}{Download taxonomizr NCBI accessions database?
Defaults to \code{TRUE}.}

\item{NCBI_accessions_name}{Name to save silva taxonomy database}

\item{silva_taxonomy_database}{Download silva taxonomy database? Defaults to
\code{TRUE}.}

\item{silva_taxonomy_name}{Name of silva headers file. Defaults
to the one hosted in MetaScope, \code{"all_silva_headers.rds"}.}

\item{blast_16S_database}{Download NCBI 16S Blast database? Defaults to
\code{TRUE}.}

\item{blast_16S_name}{Name to save NCBI 16S Blast database}
}
\description{
This function will automatically download the NCBI accessions database, the
SILVA taxonomy database, and the NCBI Blast 16S database and prepare it for
MetaScope ID and MetaBlast
}
