% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_host.R
\name{filter_host_bowtie}
\alias{filter_host_bowtie}
\title{Align reads against one or more filter libraries and subsequently
remove mapped reads}
\usage{
filter_host_bowtie(
  reads,
  unfiltered_bam,
  libs,
  lib_dir = NULL,
  output = paste(tools::file_path_sans_ext(unfiltered_bam), "filtered", "bam", sep =
    "."),
  ...,
  overwrite = FALSE
)
}
\arguments{
\item{output}{The desired name of the output .bam file. Default is
the basename of \code{reads_bam} + \code{.filtered.bam}.}

\item{...}{The optional options to pass to the bowtie align function}

\item{overwrite}{Whether existing files should be overwritten.}

\item{reads_bam}{The name of a merged, sorted .bam file that has previously
been aligned to a reference library. Likely, the output from running an
instance of \code{align_target()}.}
}
\description{
After a sample is aligned to a target library with \code{align_target_bowtie()},
we may use \code{filter_host_bowtie()} to remove unwelcome host contamination using
filter reference libraries. This function takes as input the name
of the .bam file produced via \code{align_target_bowtie()}, and produces a
sorted .bam file with any reads that match the filter libraries removed.
This resulting .bam file may be used upstream for further analysis.
}
