% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_r.R
\name{filter_r}
\alias{filter_r}
\title{Align reads against one or more filter libraries and subsequently
remove mapped reads}
\usage{
filter_r(
  reads_bam,
  libs,
  project_name = tools::file_path_sans_ext(reads_bam),
  threads = 8,
  mismatch = 5
)
}
\arguments{
\item{reads_bam}{Location of the merged and sorted .bam file to align that
was previously output by the `align_target()` function}

\item{libs}{A vector of character strings giving the basenames of the
Subread index files for alignment. These should be located in the current
directory}

\item{project_name}{A name for the project, which names the output .bam
file (e.g. project_name.bam). Defaults to the basename of the reads file.}

\item{threads}{The number of threads for the Subread alignment.
Defaults to 8}

\item{mismatch}{}
}
\value{

}
\description{
Although we have already filtered out any unmapped reads, which may belong
to one or more host species or otherwise, there may still remain some sort
of unwelcome contamination in the data from the filter species, which we
wish to remove. To do this, we employ `filter_r()`, which takes as input
the location of the BAM file created from `align_target()`, and produces a
sorted BAM file with any reads that match the filter libraries removed. We
will then use this final BAM file downstream for further analysis.
}
\examples{
# How to get previous BAM file??? Rerun code? Or include it?
# Create an index for filter libraries
# filter

}
