% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{find_taxids}
\alias{find_taxids}
\title{Gets Taxonomy IDS from accessions names
Returns a dataframe with the original accession, the associated taxonomy ID
and associated taxonomy species}
\usage{
find_taxids(all_fastas, accession_path)
}
\arguments{
\item{all_fastas}{Biostrings loaded object of all fastas}

\item{accession_path}{Path to taxonomizr accessions database}
}
\value{
Data.frame of taxonomy Ids and Species name
}
\description{
Gets Taxonomy IDS from accessions names
Returns a dataframe with the original accession, the associated taxonomy ID
and associated taxonomy species
}
\keyword{internal}
