% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_table.R
\name{generate_taxonomy_table}
\alias{generate_taxonomy_table}
\title{Generate the taxonomy table}
\usage{
generate_taxonomy_table()
}
\value{
table of taxonomic relationships for all species or strains with
NCBI genomes.
}
\description{
This code will generate the taxonomy table for all species or strains with
available NCBI genomes. Each column is a taxonomic rank or indication of a
strain; each row is a taxonomic classification for an unique species or
strain.
}
\examples{
# Code not run
\dontrun{
generate_taxonomy_table()
}
}
