% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{get_multi_seqs}
\alias{get_multi_seqs}
\title{Gets multiple sequences from different accessions in a bam file}
\usage{
get_multi_seqs(
  ids_n,
  bam_file,
  seq_info_df,
  metascope_id_tax = metascope_id_tax
)
}
\arguments{
\item{ids_n}{List of vectors with Taxonomy IDs and the number of sequences to get from each}

\item{bam_file}{A sorted bam file and index file, loaded with
Rsamtools::bamFile}
}
\value{
Biostrings format sequences
}
\description{
Returns fasta sequences from a bam file with given taxonomy IDs
}
