% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{get_seqs}
\alias{get_seqs}
\title{Gets sequences from bam file}
\usage{
get_seqs(id, bam_file, n = 10, bam_seqs)
}
\arguments{
\item{id}{Taxonomy ID of genome to get sequences from}

\item{bam_file}{A sorted bam file and index file, loaded with
Rsamtools::bamFile}

\item{n}{Number of sequences to retrieve}

\item{bam_seqs}{A list of the sequence IDs from the bam file}
}
\value{
Biostrings format sequences
}
\description{
Returns fasta sequences from a bam file with a given taxonomy ID
}
