% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaScope_ID.R
\name{locations}
\alias{locations}
\title{Helper Function for MetaScope ID}
\usage{
locations(which_taxid, which_genome, accessions, taxids, reads, bam_file)
}
\description{
Helper Function for MetaScope ID
}
