% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaScope-package.R
\docType{package}
\name{MetaScope-package}
\alias{MetaScope}
\alias{MetaScope-package}
\title{MetaScope: Tools and functions for preprocessing 16S and metagenomic sequencing microbiome data}
\description{
This package contains the tools and methods for preprocessing microbiome data. It contains tools for library generation, demultiplexing, alignment, and microbe identification. Is is essentially an R translation of PathoScope 2.0, or MetaScope.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/compbiomed/metascope https://compbiomed.github.io/metascope-docs/}
  \item Report bugs at \url{https://github.com/compbiomed/MetaScope/issues}
}

}
\author{
\strong{Maintainer}: Aubrey Odom-Mabey \email{aodom@bu.edu} (\href{https://orcid.org/0000-0001-7113-7598}{ORCID})

Authors:
\itemize{
  \item Rahul Varki \email{rvarki@bu.edu}
  \item W. Evan Johnson \email{wej@bu.edu}
}

}
\keyword{internal}
