% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_inspect.R
\name{metascope_inspect_usage}
\alias{metascope_inspect_usage}
\title{Print available arguments that can be passed to metascope_inspect()}
\usage{
metascope_inspect_usage()
}
\value{
Information about available arguments that can be passed to
metascope_inspect()
}
\description{
Calling metascope_inspect_usage() prints the available arguments
that can be passed to the ... argument of the metascope_inspect() function of
the package. Note that some arguments are invalid if they are already
handled as explicit function arguments.
}
\examples{
metascope_inspect_usage()
}
\references{
{
Langmead B, Wilks C, Antonescu V, Charles R. Scaling
read aligners to hundreds of threads on general-purpose processors.
Bioinformatics. 2018 Jul 18. doi: 10.1093/bioinformatics/bty648.

Langmead B, Salzberg SL. Fast gapped-read alignment with
Bowtie 2. Nature Methods. 2012 Mar 4;9(4):357-9. doi: 10.1038/nmeth.1923.
}
}
