% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_host.R
\name{mk_interim_fastq}
\alias{mk_interim_fastq}
\title{Helper function for filter_host_bowtie to write interim fastq file}
\usage{
mk_interim_fastq(bf, read_loc, maxMemory)
}
\description{
Helper function for filter_host_bowtie to write interim fastq file
}
