% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_r.R
\name{remove_matches}
\alias{remove_matches}
\title{Helper function to remove reads matched to filter libraries}
\usage{
remove_matches()
}
\description{
Helper function to remove reads matched to filter libraries
}
