% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_headers}
\alias{species_headers}
\title{An object containing species identifications for SILVA}
\format{
tibble
}
\usage{
data(species_headers)
}
\description{
This \code{data.frame} contains taxonomy IDs along with genus and
species-level assignments for the SILVA 13_8 database.
}
\examples{
data("species_headers")

}
\keyword{datasets}
