% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{taxid_to_name}
\alias{taxid_to_name}
\title{Converts NCBI taxonomy ID to scientific name}
\usage{
taxid_to_name(taxids, NCBI_key = NULL)
}
\arguments{
\item{taxids}{List of NCBI taxids to convert to scientific name}

\item{NCBI_key}{(character) NCBI Entrez API key. optional. See
taxize::use_entrez(). Due to the high number of requests made to NCBI, this
function will be less prone to errors if you obtain an NCBI key.}
}
\value{
Returns a dataframe of blast results for a metascope result
}
\description{
Converts NCBI taxonomy ID to scientific name
}
