% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_table}
\alias{taxonomy_table}
\title{Taxonomy table}
\format{
An object of class \code{data.frame} with 74616 rows and 9 columns.
}
\usage{
taxonomy_table
}
\description{
This is a taxonomy table to be used in \code{download_refseq()}. Last updated 10/18/22.
}
\examples{
data("taxonomy_table")

}
\keyword{datasets}
