% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaScope_ID.R
\name{unique_identifier}
\alias{unique_identifier}
\title{Determines if a read is multi-mapping}
\usage{
unique_identifier(x)
}
\arguments{
\item{x}{Integer. An integer representing the group size of a specific read.}
}
\value{
either 0 (unique) or 1 (multi-mapping)
}
\description{
A read should be multi-mapping if it appears more than once in a bam file. 
The function accepts the read group size which will either be 1 if the read
is mapped uniquely or >1 if the read is multi-mapping. The function will 
return 0 if the read is unique and 1 if the read is multi-mapping. 
The function is intended to be used to create the uniqueness indicator 
(y_ind_2) which is used for calculating theta. Not meant to be used outside
of function.
}
