% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_host.R
\name{write_chunks}
\alias{write_chunks}
\title{Helper function for filter_host_bowtie to write chunks into fastq}
\usage{
write_chunks(bf, to)
}
\description{
Helper function for filter_host_bowtie to write chunks into fastq
}
