% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{add_in_taxa_ncbi}
\alias{add_in_taxa_ncbi}
\title{Adds in taxa if NCBI database
Returns MetaScope Table with NCBI taxa in separate columns}
\usage{
add_in_taxa_ncbi(metascope_id_in, accession, param)
}
\arguments{
\item{metascope_id_in}{MetaScope ID file with NCBI taxa qnames}

\item{accession_path}{(character) Path to accession.sql file.}
}
\description{
Adds in taxa if NCBI database
Returns MetaScope Table with NCBI taxa in separate columns
}
\keyword{internal}
