% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_target.R
\name{align_target}
\alias{align_target}
\title{Align microbiome reads to a set of reference libraries}
\usage{
align_target(
  reads,
  libs,
  lib_dir = NULL,
  project_name = tools::file_path_sans_ext(reads),
  settings = align_details
)
}
\arguments{
\item{reads}{Location of the .fastq file to align}

\item{libs}{A vector of character strings giving the basenames of the
Subread index files for alignment. These should be located in the same
directory and the current directory if lib_dir=NULL.}

\item{lib_dir}{path to the library index files (all libraries should be here)}

\item{project_name}{A name for the project, which names the output .bam
file (e.g. project_name.bam). Defaults to the basename of the reads file.}

\item{settings}{A named \code{list} specifying alignment parameters for
the \code{Rsubread::align()} function, which is
called inside \code{align_target()}.
Elements should include type, nthreads, maxMismatches, nsubreads,
phredOffset, unique, and nBestLocations. Descriptions of these parameters
are available under \code{?Rsubread::align}. Defaults to the global
\code{align_details} object.}
}
\value{
This function writes a merged and sorted .bam file after aligning to all
reference libraries given, along with a summary report file, to the user's
working directory.
The function also outputs the new .bam filename.
}
\description{
This is the main MetaScope target library mapping function, using Rsubread
and multiple libraries. Aligns to each library separately, filters
unmapped reads from each file, and then merges and sorts the .bam files
from each library into one output file. If desired, output can be
passed to `filter_host()` to remove reads that also map to filter library
genomes.
}
\examples{

## Create object with path to example reference fasta file
refPath <- system.file("extdata","Mononegavirales.fasta", package = "MetaScope")

## Copy the example reference fasta file to the current directory
file.copy(from = refPath, to = file.path(".", "Mononegavirales.fasta"))

## Make and align to a single reference genome library
mk_subread_index("Mononegavirales.fasta")
readPath <- system.file("extdata", "virus_example.fastq",package = "MetaScope")
viral_map <- align_target(readPath, "Mononegavirales", project_name="virus_example")
viral_map_sam <- Rsamtools::asSam(viral_map, overwrite = TRUE)

## Make and align to multiple reference genome libraries
mk_subread_index('Mononegavirales.fasta', split = 0.002)
targLibs <- c("Mononegavirales_1", "Mononegavirales_2")
readPath <- system.file("extdata", "virus_example.fastq",package = "MetaScope")
viral_map <- align_target(readPath, targLibs, project_name = "virus_example")

}
