% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_target_bowtie.R
\name{align_target_bowtie}
\alias{align_target_bowtie}
\title{Align microbiome reads to set of indexed Bowtie2 libraries}
\usage{
align_target_bowtie(
  read1,
  read2 = NULL,
  lib_dir,
  libs,
  align_dir,
  align_file,
  align_format = "bam",
  bowtie2_options = NULL,
  threads = 8,
  overwrite = FALSE
)
}
\arguments{
\item{read1}{Path to the .fastq file to align.}

\item{read2}{Optional: Location of the mate pair .fastq file to align.}

\item{lib_dir}{Path to the directory that contains the Bowtie2 indexes.}

\item{libs}{The basename of the Bowtie2 indexes to align against 
(without trailing .bt2 or .bt2l extensions).}

\item{align_dir}{Path to the directory where the output alignment file 
should be created.}

\item{align_file}{The basename of the output alignment file file 
(without trailing .sam or .bam extensions).}

\item{align_format}{The format of the alignment file. Default is "bam" 
but can also pass "sam" to the function. NOTE: To create the bam file, first
the sam file must be created but afterwards the sam file will be deleted.}

\item{bowtie2_options}{Optional: Additional parameters that can be passed to
the align_target_bowtie() function. To see all the available parameters
use Rbowtie2::bowtie2_usage(). Default parameters are the parameters are the 
default parameters that PathoScope 2.0 uses. NOTE: Users should pass all their
parameters as one string and if optional parameters are given then the user 
is responsible for entering all the parameters to be used by Bowtie2. NOTE:
The only parameters that should NOT be specified here is the threads.}

\item{threads}{The number of threads that can be utilized by the function.
Default is 8 threads.}

\item{overwrite}{Whether existing files should be overwritten. 
Default is FALSE.}
}
\value{
Returns the path to the directory where the output alignment 
file is stored.
}
\description{
Align microbiome reads to set of indexed Bowtie2 libraries
}
\examples{
# Code not run
\donttest{

## Create alignment bam file produced by Bowtie2

# Create a temporary directory to store reference fasta files
ref_temp <- tempfile()
dir.create(ref_temp)

# Create a temporary directory to store the index files 
lib_temp <- tempfile()
dir.create(lib_temp)

# Create a temporary directory to store the index files 
align_temp <- tempfile()
dir.create(align_temp)

# Download all the RefSeq reference viral genomes to current directory
download_refseq('viruses', compress = TRUE)

# Move downloaded fasta file to temporary reference directory 
file.rename(from = file.path(".", "Viruses.fasta.gz"), 
to = file.path(ref_temp, "Viruses.fasta.gz"))

# Create bowtie index files in temporary index directory
mk_bowtie_index(ref_dir = ref_temp, lib_dir = lib_temp, lib_name = "virus", 
overwrite=FALSE)

# Get path to example reads
readPath <- system.file("extdata", "virus_example.fastq", package = "MetaScope")

# Create alignment file 
align_target_bowtie(read1 = readPath, lib_dir = lib_temp, 
libs = "virus", align_dir = align_temp, align_file = "example", 
align_format = "bam", overwrite = TRUE)
}
}
