% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{blast_reassignment}
\alias{blast_reassignment}
\title{Reassign reads from Blast alignment}
\usage{
blast_reassignment(
  metascope_blast_path,
  species_threshold,
  num_hits,
  blast_tmp_dir,
  out_dir,
  sample_name
)
}
\arguments{
\item{species_threshold}{species threshold for validation
Species percentage hit (Blasted reads) threshold to validate a taxa
Threshold >0.5 to be considered  (50%, so this is a proportion, can be 0.2 default)}

\item{num_hits}{number of hits}

\item{blast_tmp_dir}{directory with blast results}

\item{out_dir}{Character string, path to output directory.}

\item{sample_name}{Character string, sample name for output files.}

\item{metascope_blast_df}{Path to \code{metascope_blast} CSV output file.}
}
\value{
Returns an updated blast reassignment file
}
\description{
This function allows the user to reassign reads who's NCBI BLAST results
contradict results provided by MetaScope to assignments that were BLAST
validated. It returns an updated csv with reads reassigned according to their
BLAST validation.
}
