% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_animalcules.R
\name{convert_animalcules}
\alias{convert_animalcules}
\title{Create a multi-assay experiment from MetaScope output for usage with animalcules
Upon completion of the MetaScope pipeline, users can analyze and visualize
abundances in their samples using the animalcules package. This function
allows interoperability of metascope_id output with both animalcules and
QIIME (coming soon!).}
\usage{
convert_animalcules(
  meta_counts,
  annot_path,
  which_annot_col,
  end_string = ".filtered.metascope_id.csv",
  qiime_biom_out = FALSE
)
}
\arguments{
\item{meta_counts}{A vector of filepaths to the counts ID CSVs output
by MetaScope}

\item{annot_path}{The filepath to the annotation file for the samples}

\item{which_annot_col}{The name of the column of the annotation file
containing the sample IDs.
These should be the same as the \code{meta_counts} root filenames.}

\item{end_string}{The end string used at the end of the metascope_id
files. Default is ".filtered.metascope_id.csv"}

\item{qiime_biom_out}{Would you also like a qiime-compatible biom file
output? Default is \code{FALSE}.}
}
\value{
returns a multi-assay experiment file of combined sample counts
data and/or biom file for QIIME. The multi-assay experiment will have
assays for the counts ("MGX"), log counts, CPM, and log CPM.
}
\description{
Create a multi-assay experiment from MetaScope output for usage with animalcules
Upon completion of the MetaScope pipeline, users can analyze and visualize
abundances in their samples using the animalcules package. This function
allows interoperability of metascope_id output with both animalcules and
QIIME (coming soon!).
}
\examples{
\donotrun{
all_files <- list.files("~/decamp/analysis/aodom/Novartis_COPD/MetaScope_run/Output", pattern = "*.filtered.metascope_id.csv", full.names = TRUE)
out <- convert_animalcules(meta_counts = all_files, annot_path = "~/decamp/analysis/aodom/Novartis_COPD/Data/DECAMP_patho_annotation.csv", end_string = ".filtered.metascope_id.csv", which_annot_col = "DECAMP_ID")
}
 # coming soon - this is from Aubrey
 
}
