% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_host.R
\name{filter_host}
\alias{filter_host}
\title{Align reads against one or more filter libraries and subsequently
remove mapped reads}
\usage{
filter_host(
  reads_bam,
  libs,
  output = paste(tools::file_path_sans_ext(reads_bam), "filtered", "bam", sep = "."),
  settings = align_details
)
}
\arguments{
\item{reads_bam}{The name of a merged, sorted .bam file that has previously
been aligned to a reference library. Likely, the output from running an
instance of \code{align_target()}.}

\item{libs}{A vector of character strings giving the basenames of the
Subread index files for alignment. These should be located in the current
directory.}

\item{output}{The desired name of the output .bam file. Default is
the basename of \code{reads_bam} + \code{.filtered.bam}.}

\item{settings}{A named \code{list} specifying alignment parameters for
the \code{Rsubread::align()} function, which is
called inside \code{align_target()}.
Elements should include type, nthreads, maxMismatches, nsubreads,
phredOffset, unique, and nBestLocations. Descriptions of these parameters
are available under \code{?Rsubread::align}. Defaults to the global
\code{align_details} object.}
}
\value{
The name of a filtered, sorted .bam file written to the user's
current working directory.
}
\description{

}
\examples{
# How to get previous BAM file??? Rerun code? Or include it?
# Create an index for filter libraries
# filter

}
