% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_id.R
\name{locations}
\alias{locations}
\title{Helper Function for MetaScope ID}
\usage{
locations(
  which_taxid,
  which_genome,
  accessions,
  taxids,
  reads,
  out_base,
  out_dir
)
}
\arguments{
\item{which_taxid}{Which taxid to plot}

\item{which_genome}{Which genome to plot}

\item{accessions}{List of accessions from \code{metascope_id()}}

\item{taxids}{List of accessions from \code{metascope_id()}}

\item{reads}{List of reads from input file}

\item{input_file}{The path to the input file}
}
\value{
A plot of the read coverage for a given genome
}
\description{
Used to create plots of genome coverage for any number of accession numbers
}
