% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaScope-package.R
\docType{package}
\name{MetaScope-package}
\alias{MetaScope}
\alias{MetaScope-package}
\title{MetaScope: Tools and functions for preprocessing 16S and metagenomic sequencing microbiome data}
\description{
This package contains tools and methods for preprocessing microbiome data. Functionality includes library generation, demultiplexing, alignment, and microbe identification. It is in part an R translation of the PathoScope 2.0 pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wejlab/metascope https://wejlab.github.io/metascope-docs/}
  \item Report bugs at \url{https://github.com/wejlab/MetaScope/issues}
}

}
\author{
\strong{Maintainer}: Aubrey Odom \email{aodom@bu.edu} (\href{https://orcid.org/0000-0001-7113-7598}{ORCID})

Authors:
\itemize{
  \item Sean Lu \email{sl1729@njms.rutgers.edu} (\href{https://orcid.org/0009-0007-8005-6125}{ORCID})
  \item Rahul Varki \email{rvarki@bu.edu} (\href{https://orcid.org/0009-0003-5721-9484}{ORCID})
  \item W. Evan Johnson \email{wej@bu.edu} (\href{https://orcid.org/0000-0002-6247-6595}{ORCID})
}

}
\keyword{internal}
