% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_bowtie_index.R
\name{mk_bowtie_index}
\alias{mk_bowtie_index}
\title{Make a Bowtie2 index}
\usage{
mk_bowtie_index(
  ref_dir,
  index_dir,
  index_name = basename(index_dir),
  threads = 4,
  overwrite = FALSE
)
}
\arguments{
\item{ref_dir}{The name/location of the directory that contains the reference
libraries in uncompressed fasta format}

\item{index_dir}{The name/location of the directory where the index files 
should be created}

\item{index_name}{The basename of the index files. Default is the name of 
the last directory specified in index_dir}

\item{threads}{The number of threads to be used. Default is 4.}

\item{overwrite}{Whether existing files should be overwritten. Default is false.}
}
\value{
Creates multiple large Bowtie2 indexes (.bt2l) for the supplied 
reference .fasta files. Returns the file path of the folder holding these
files.
}
\description{
This function is a wrapper for the \code{Rbowtie2::bowtie2_build} function.
It will create large Bowtie2 indexes (.bt2l) from a reference fasta files.
}
