% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_bowtie_index.R
\name{mk_bowtie_index}
\alias{mk_bowtie_index}
\title{Make a Bowtie2 index}
\usage{
mk_bowtie_index(
  ref_dir,
  index_dir,
  index_name = basename(index_dir),
  threads = 4,
  overwrite = FALSE
)
}
\arguments{
\item{ref_dir}{\code{Character} scalar. The name/location of the directory that contains the reference
libraries in uncompressed fasta format}

\item{index_dir}{\code{Character} scalar. The name/location of the directory where the index files 
should be created}

\item{index_name}{\code{Character} scalar. The basename of the index files to be created. Default is the
basename of the index_dir}

\item{threads}{\code{Integer}. The number of threads to be used. Default is 4.}

\item{overwrite}{\code{Logical}. Whether existing files should be overwritten. Default is FALSE.}
}
\value{
Creates multiple Bowtie2 indexes (.bt2 or .bt2l) of the supplied 
reference .fasta files. Returns the file path of the folder containing these
files.
}
\description{
This function is a wrapper for the \code{Rbowtie2::bowtie2_build} function.
It will create either small (.bt2) or large Bowtie2 indexes (.bt2l) of the 
given reference fasta files depending on their combined file size.
}
