% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{rBLAST_single_result}
\alias{rBLAST_single_result}
\title{rBLAST_single_result}
\usage{
rBLAST_single_result(
  results_table,
  bam_file,
  which_result,
  num_reads = 100,
  hit_list = 10,
  num_threads = 1,
  db_path,
  quiet = quiet,
  NCBI_key = NULL,
  bam_seqs
)
}
\arguments{
\item{results_table}{A dataframe of the Metascope results}

\item{bam_file}{A sorted bam file and index file, loaded with
Rsamtools::bamFile}

\item{which_result}{Index in results_table for which result to Blast search}

\item{num_reads}{Number of reads to blast per result}

\item{hit_list}{Number of how many blast results to fetch per read}

\item{num_threads}{Number of threads if multithreading}

\item{db_path}{Blast database path}

\item{quiet}{Logical, whether to print out more informative messages. Default
is FALSE.}

\item{NCBI_key}{(character) NCBI Entrez API key. optional. See
taxize::use_entrez(). Due to the high number of requests made to NCBI, this
function will be less prone to errors if you obtain an NCBI key.}

\item{bam_seqs}{A list of the sequence IDs from the bam file}
}
\value{
Returns a dataframe of blast results for a metascope result
}
\description{
rBLAST_single_result
}
