% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_host.R
\name{remove_matches}
\alias{remove_matches}
\title{Helper function to remove reads matched to filter libraries}
\usage{
remove_matches(reads_bam, read_names, name_out)
}
\arguments{
\item{reads_bam}{The name of a merged, sorted .bam file that has previously
been aligned to a reference library. Likely, the output from running an
instance of `align_target()`.}

\item{read_names}{A `list` of target query names from `reads_bam` that
have also aligned to a filter reference library. Each `list` element
should be a vector of read names.}

\item{name_out}{The name of the .bam file that to which the filtered alignments
will be written.}
}
\value{
The name of a filtered, sorted .bam file written to the user's
current working directory.
}
\description{
Within the `filter_host()` function, we align our sequencing sample to all
filter libraries of interest. The `remove_matches()` function allows
for removal of any target reads that are also aligned to filter libraries.
}
