% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metascope_blast.R
\name{taxid_to_name}
\alias{taxid_to_name}
\title{Converts NCBI taxonomy ID to scientific name}
\usage{
taxid_to_name(taxids, accession_path)
}
\arguments{
\item{taxids}{List of NCBI taxids to convert to scientific name}

\item{accession_path}{(character) Filepath to NCBI accessions SQL
database. See \code{taxonomzr::prepareDatabase()}.}
}
\value{
Returns a dataframe of blast results for a metascope result
}
\description{
Converts NCBI taxonomy ID to scientific name
}
